refactor: move search to feature folder
