+++ packages/api-reference/src/features/Search/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchButton } from './SearchButton.vue'
